+++ src/GNUmakefile
@@ -55,6 +55,7 @@ OBJECTS += \
 	extension/extension.o	\
 	extension/kompat/kompat.o \
 	extension/fake_id0/chown.o \
+	extension/fake_id0/sendmsg.o \
 	extension/fake_id0/fake_id0.o \
 	extension/link2symlink/link2symlink.o \
 	loader/loader-wrapped.o
